requirements.txt
Flask
pandas
langchain
langchain-groq 
Jinja2 
python-dotenv
requests 
